+++ src/index.ts
@@ -1,4 +1,19 @@
+import {
+  HandlebarsConfigurationError,
+  HandlebarsError,
+  HandlebarsInvalidPathError,
+  HandlebarsRenderError,
+  HandlebarsTemplateNotFoundError,
+} from './handlebars.error';
 import { HandlebarsModule } from './handlebars.module';
 import { HandlebarsService } from './handlebars.service';
 
-export { HandlebarsModule, HandlebarsService };
+export {
+  HandlebarsConfigurationError,
+  HandlebarsError,
+  HandlebarsInvalidPathError,
+  HandlebarsModule,
+  HandlebarsRenderError,
+  HandlebarsService,
+  HandlebarsTemplateNotFoundError,
+};